# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "opendataAPI",
        "description": "Open data API from open dataset",
        "contact": {
            "email": "mellyniaramadhan762@gmail.com"
        },
        "license": {
            "name": "mellyniaramadhan",
            "url": "http://your-license-url.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/BPS/RiceProductionByDistricts": {
            "get": {
                "tags": [
                    "Rice Production"
                ],
                "summary": "Get a paginated list of rice production data by districts.",
                "operationId": "ef72c4aad1899a63f887670ce505e7f2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RiceProductionByDistricts"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/BPS": {
            "post": {
                "tags": [
                    "Rice Production"
                ],
                "summary": "Create a new rice production record.",
                "operationId": "4aea0bb546e11967ab15c4f4e8b599d0",
                "requestBody": {
                    "description": "Data to be added",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "CodeProvince": {
                                        "type": "integer"
                                    },
                                    "ProvinceName": {
                                        "type": "string"
                                    },
                                    "CodeDistrict": {
                                        "type": "integer"
                                    },
                                    "DistrictName": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "CodeProvince": "example CodeProvince",
                                "ProvinceName": "example ProvinceName",
                                "CodeDistrict": "example CodeDistrict",
                                "DistrictName": "Example DistrictName",
                                "TonsIn2020": "Example TonsIn2020",
                                "TonsIn2021": "Example TonsIn2021",
                                "TonsIn2022": "Example TonsIn2022",
                                "TonsIn2023": "Example TonsIn2023"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "CodeProvince": {
                                            "type": "integer",
                                            "example": "32"
                                        },
                                        "ProvinceName": {
                                            "type": "string",
                                            "example": "Jawa Barat"
                                        },
                                        "CodeDistricts": {
                                            "type": "integer",
                                            "example": 3201
                                        },
                                        "DistrictsName": {
                                            "type": "string",
                                            "example": "Bogor"
                                        },
                                        "TonsIn2020": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "TonsIn2021": {
                                            "type": "integer",
                                            "example": 28515
                                        },
                                        "TonsIn2022": {
                                            "type": "integer",
                                            "example": 299893.8
                                        },
                                        "TonsIn2023": {
                                            "type": "integer",
                                            "example": 281153
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "fieldName": [
                                                    "Error message"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/BPS/RiceProductionByDistricts/{id}": {
            "get": {
                "tags": [
                    "Rice Production"
                ],
                "summary": "Get a specific rice production record by ID.",
                "description": "Display the specified resource.",
                "operationId": "20656617a1f16dab2085e0effdcefd57",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the rice production record to be retrieved.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/RiceProductionByDistricts"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Message": {
                                            "type": "string",
                                            "example": "Not Found!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rice Production"
                ],
                "summary": "Delete a rice production record by ID.",
                "operationId": "8d1672d5c8df55584e2678c688d8dd0a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the rice production record to be deleted.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Message": {
                                            "type": "string",
                                            "example": "Deleted Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "RiceProductionByDistricts": {
                "properties": {
                    "CodeProvince": {
                        "type": "string"
                    },
                    "ProvinceName": {
                        "type": "string"
                    },
                    "CodeDistricts": {
                        "type": "integer"
                    },
                    "DistrictsName": {
                        "type": "string"
                    },
                    "TonsIn2020": {
                        "type": "integer"
                    },
                    "TonsIn2021": {
                        "type": "integer"
                    },
                    "TonsIn2022": {
                        "type": "integer"
                    },
                    "TonsIn2023": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        }
    }
}